+++ examples/countdown/src/main.rs
@@ -1,16 +1,16 @@
 //! Demonstrates the TIMERS service, variable mapping and profiler trace recording.
 
-use tracing_subscriber::{layer::SubscriberExt as _, util::SubscriberInitExt as _};
 use zng::{image, prelude::*, widget::background_color};
 
 fn main() {
+    unsafe {
+        // see zng::app::trace_recorder for more details
+        std::env::set_var("ZNG_RECORD_TRACE", "");
+        std::env::set_var("ZNG_RECORD_TRACE_FILTER", "debug");
+    }
     zng::env::init!();
 
-    // record profiler, use chrome://tracing or <ui.perfetto.dev>.
-    let (chrome_layer, _guard) = tracing_chrome::ChromeLayerBuilder::new().include_args(true).build();
-    tracing_subscriber::registry().with(chrome_layer).init();
-
-    zng::view_process::default::run_same_process(app_main);
+    app_main();
 }
 
 fn app_main() {